backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pyjwt==2.8.0
cryptography==41.0.7
requests==2.31.0
